Rename ReasonReact module to just React
