+++ lib/js/src/ReactOptimizedCreateClass.js
